# Lazy initialize hashes