--- a/cng/hmac.go
+++ b/cng/hmac.go
@@ -12,26 +12,6 @@ import (
 	"github.com/microsoft/go-crypto-winnative/internal/bcrypt"
 )
 
-// hashToID converts a hash.Hash implementation from this package
-// to a CNG hash ID
-func hashToID(h hash.Hash) string {
-	if _, ok := h.(*hashX); !ok {
-		return ""
-	}
-	var id string
-	switch h.Size() {
-	case 20:
-		id = bcrypt.SHA1_ALGORITHM
-	case 256 / 8:
-		id = bcrypt.SHA256_ALGORITHM
-	case 384 / 8:
-		id = bcrypt.SHA384_ALGORITHM
-	case 512 / 8:
-		id = bcrypt.SHA512_ALGORITHM
-	}
-	return id
-}
-
 // NewHMAC returns a new HMAC using BCrypt.
 // The function h must return a hash implemented by
 // CNG (for example, h could be cng.NewSHA256).